Since the header length can't be greater than ushort, same can apply for ContentTypeLength

+++ MicroProtocol/Headers/ContentHeader.cs
@@ -10,7 +10,7 @@ namespace Ace.Networking.MicroProtocol.Headers
         }
 
         public byte[] ContentType { get; set; }
-        public int ContentTypeLength => ContentType?.Length ?? 0;
+        public ushort ContentTypeLength => (ushort)(ContentType?.Length ?? 0);
 
         public int ContentLength { get; set; }
 
@@ -18,8 +18,8 @@ namespace Ace.Networking.MicroProtocol.Headers
         public override BasicHeader Deserialize(byte[] target, int offset = 0)
         {
             base.Deserialize(target, offset);
-            var contentTypeLength = BitConverter.ToInt32(target, offset + Position);
-            Position += sizeof(int);
+            var contentTypeLength = BitConverter.ToUInt16(target, offset + Position);
+            Position += sizeof(ushort);
             ContentType = new byte[contentTypeLength];
             for (var i = 0; i < contentTypeLength; i++)
             {
@@ -34,8 +34,8 @@ namespace Ace.Networking.MicroProtocol.Headers
         public override void Serialize(byte[] target, int offset = 0)
         {
             base.Serialize(target, offset);
-            BitConverter2.GetBytes(ContentTypeLength, target, offset + Position);
-            Position += sizeof(int);
+            BitConverter2.GetBytes((short)ContentTypeLength, target, offset + Position);
+            Position += sizeof(ushort);
             //Encoding.ASCII.GetBytes(ContentType, 0, ContentTypeLength, target, offset + Position); Position += ContentTypeLength;
             for (var i = 0; i < ContentTypeLength; i++)
             {
